+++ python/automation-api/README.md
@@ -1,17 +1,16 @@
 # Python Automation API Example
 Provides an example of using automation API using Python.
 
-
 ## Main Talking Points
 * Shows the automation API SDK in action.
   * Show how you are basically doing `pulumi up` and `pulumi destroy`, etc in code instead of command line.
 * This code can then be plugged into a web-service or similar system to enable self-service, etc.
-* Although the `deployments.json` file is hand-crafted in this example, it could be generated by parsing the folders containing the Pulumi projects.
+* Although the `arrangements.yaml` file is hand-crafted in this example, it could be generated by parsing the folders containing the Pulumi projects.
 
 ## Main Components of the Example
 
 * `arrangements.yaml`: This file captures the information needed for the orchestration logic. Specifically, it provides information on how to find and run specific stacks including multi-stack use-cases. It is hand-crafted to allow you to demo what you want. 
-* `orchestration`: This folder contains the orchestration logic used to manage the projects and stacks. It is called by the code in the `run-automation` folder.
+* `utils`: This folder contains various utilities including the orchestration logic used to deploy the stacks. It is called by the `automation.py`.
 * `automation.py`: This is the main program that executes the orchestration automation.
 
 ## Running the Example
