+++ python/automation-api/arrangements.yaml
@@ -3,17 +3,17 @@
 # If an arrangement consists of more than one project, then the projects need to be listed in the order they 
 # need to deployed. (Note: the logic will reverse the order for destroy operations.)
 
-# Absolute or relative folder for the automation code to use 
-# with respect to the folders provided for the arrangement. 
+# Absolute or relative path to where the projects are located.
+# Relative path must be relative to where the `automation.py` program is run.
 base-folder: ../.. 
 
 # Array of arrangements and the related project folder(s) for each arrangement.
 arrangements:
 - name: AWS Serverless
-  projects-folders:
+  project-folders:
     - typescript/aws-ts-serverless
-# - name: Guestbook 
-#   project-folders:
-#     - csharp/aws-cs-eks-guestbook/eks-base-infra
-#     - csharp/aws-cs-eks-guestbook/guestbook-app
-#     - yaml/aws-yaml-eks-guestbook/guestbook-app-yaml
+- name: Guestbook 
+  project-folders:
+    - csharp/aws-cs-eks-guestbook/eks-base-infra
+    - csharp/aws-cs-eks-guestbook/guestbook-app
+    - yaml/aws-yaml-eks-guestbook/guestbook-app-yaml
