+++ python/automation-api/utils/stack.py
